introduce ScriptCheck class and improve testing
