+++ anaconda_linter/lint/check_build_help.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 import os
 import re
 from pathlib import Path
-from typing import Any, Final, Optional
+from typing import Final
 
 from conda.models.match_spec import MatchSpec
 from conda_recipe_manager.parser.dependency import Dependency, DependencySection
@@ -18,7 +18,7 @@ from conda_recipe_manager.parser.recipe_reader_deps import RecipeReaderDeps
 from percy.render.recipe import Recipe
 
 from anaconda_linter import utils as _utils
-from anaconda_linter.lint import LintCheck, Severity
+from anaconda_linter.lint import LintCheck, ScriptCheck, Severity
 
 # Does not include m2-tools, which should be checked using wild cards.
 BUILD_TOOLS: Final[tuple] = (
@@ -496,7 +496,7 @@ class uses_setup_py(LintCheck):
         return recipe.patch(op)
 
 
-class pip_install_args(LintCheck):
+class pip_install_args(ScriptCheck):
     """
     `pip install` should be run with --no-deps and --no-build-isolation.
 
@@ -506,88 +506,18 @@ class pip_install_args(LintCheck):
 
     """
 
-    @staticmethod
-    def _check_line(line: str) -> bool:
-        """
-        Check a line for a broken call to pip install
-        """
+    def _check_line(self, line: str) -> bool:
         if "pip install" in line:
             required_args = ["--no-deps", "--no-build-isolation"]
             if any(arg not in line for arg in required_args):
                 return True
         return False
 
-    def _check_block(self, path: str, value: Any) -> str:
-        """
-        Check a line or a list of lines for a broken call to pip install
-        """
-        if value is None:
-            return ""
-        if isinstance(value, str):
-            if not self._check_line(value):
-                return ""
-            return path
-        for idx, line in enumerate(value):
-            if self._check_line(line):
-                return path + "/" + str(idx)
-
-    def _check_build_sh(self, recipe_dir: Optional[Path], build_script: Any) -> None:
-        """
-        Check a build.sh file for a broken call to pip install
-        """
-        if build_script:
-            if isinstance(build_script, list):
-                return
-            build_file = recipe_dir / build_script
-        else:
-            build_file = recipe_dir / "build.sh"
-        if build_file.exists():
-            with open(build_file, mode="r", encoding="utf-8") as build_sh:
-                for line in build_sh:
-                    if self._check_line(line):
-                        self.message(section=f"build script: {str(build_file)}")
-
     def check_recipe(self, recipe_name: str, arch_name: str, recipe: RecipeReaderDeps) -> None:
-        recipe_dir: Final[Optional[Path]] = Path(recipe_name) if recipe_name else None
-        # Check root level
-        build_script_path: Final[str] = "/build/script"
-        build_script_val = None
-        if recipe.contains_value(build_script_path):
-            build_script_val = recipe.get_value(build_script_path)
-            if build_script_path := self._check_block(build_script_path, build_script_val):
-                self.message(section=build_script_path)
-        if recipe_dir:
-            self._check_build_sh(recipe_dir, build_script_val)
-        # Check outputs
-        for package in recipe.get_package_paths():
-            if package == "/":
-                continue
-            script_path = recipe.append_to_path(package, "/script")
-            output_build_script_path = recipe.append_to_path(package, "/build/script")
-            script_val = None
-            if recipe.contains_value(script_path):
-                script_val = recipe.get_value(script_path)
-                if script_path := self._check_block(script_path, script_val):
-                    self.message(section=script_path)
-            elif recipe.contains_value(output_build_script_path):
-                script_val = recipe.get_value(output_build_script_path)
-                if output_build_script_path := self._check_block(output_build_script_path, script_val):
-                    self.message(section=output_build_script_path)
-            if recipe_dir:
-                self._check_build_sh(recipe_dir, script_val)
+        self._check_build_script(recipe_name, recipe)
 
     def fix(self, message, data) -> bool:
-        section = message.section
-        if section.startswith("build script: "):
-            return False
-        recipe = self.unrendered_recipe
-        return recipe.patch(
-            {
-                "op": "replace",
-                "path": section,
-                "value": "{{ PYTHON }} -m pip install . --no-deps --no-build-isolation",
-            }
-        )
+        return self._fix_script(message)
 
 
 class python_build_tools_in_host(LintCheck):
