+++ tests/test_aux_files/auto_fix/pip_install_args.yaml
@@ -25,7 +25,9 @@ outputs:
       host:
         - pip
   - name: output3
-    script: {{ PYTHON }} -m pip install . --no-deps
+    script:
+      - dummy_command
+      - {{ PYTHON }} -m pip install . --no-deps
     requirements:
       host:
         - pip
